Added a new wish list page for danmaku feature. Testing this angle from consumer side.
