+++ frontend/danmaku_main.ts
@@ -1,0 +1,7 @@
+import { DanmakuBody } from "./danmaku_body";
+
+async function main(): Promise<void> {
+  DanmakuBody.create();
+}
+
+main();
