+++ frontend/fan_main.ts
@@ -1,0 +1,7 @@
+import { FanBody } from "./fan_body";
+
+async function main(): Promise<void> {
+  FanBody.create();
+}
+
+main();
