+++ frontend/sec_main.ts
@@ -1,0 +1,7 @@
+import { SecBody } from "./sec_body";
+
+async function main(): Promise<void> {
+  SecBody.create();
+}
+
+main();
