+++ package.json
@@ -44,6 +44,7 @@
     "postmark": "^4.0.5"
   },
   "assetExts": [
-    ".png"
+    ".png",
+    ".mp4"
   ]
 }
