+++ prod/frontend_danmaku_main.ts
@@ -1,0 +1,2 @@
+import "./env";
+import "../frontend/danmaku_main";
