+++ service_interface/client.ts
@@ -1,4 +1,4 @@
-import { JoinWaitListRequestBody, JoinWaitListResponse, JOIN_WAIT_LIST } from './interface';
+import { JoinWaitListRequestBody, JoinWaitListResponse, JOIN_WAIT_LIST, WishlistRequestBody, WishlistResponse, WISHLIST } from './interface';
 import { ClientRequestInterface } from '@selfage/service_descriptor/client_request_interface';
 
 export function newJoinWaitListRequest(
@@ -6,6 +6,15 @@ export function newJoinWaitListRequest(
 ): ClientRequestInterface<JoinWaitListResponse> {
   return {
     descriptor: JOIN_WAIT_LIST,
+    body,
+  };
+}
+
+export function newWishlistRequest(
+  body: WishlistRequestBody,
+): ClientRequestInterface<WishlistResponse> {
+  return {
+    descriptor: WISHLIST,
     body,
   };
 }
